refactor integration test code.
