+++ .github/workflows/integration_test.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches: [ "*" ]
   pull_request:
-    branches: [ "*" ]
+    branches: [ "main" ]
   schedule:
     - cron: '0 */12 * * *'
 jobs:
