+++ src/tests/integrationTest/IntegrationTester/Models/BalanceModel.cs
@@ -1,0 +1,9 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+public class BalanceModel
+{
+    public string UserName { get; set; }
+    public int Balance { get; set; }
+}
+
