+++ src/tests/integrationTest/IntegrationTester/Models/CountorModel.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+public class CountorModel
+{
+    public int StartValue { get; set; }
+    public int CurrentSum { get; set; }
+    public int BatchExecutedCount { get; set; }
+    public int TotalCount { get; set; }
+}
+
