+++ src/tests/integrationTest/IntegrationTester/Utility/RabbitMqSetting.cs
